# reference/openapi.yaml
openapi: 3.1.0
info:
  title: Technical Terms API
  description: "A searchable glossary of technical terminology used in documentation\
    \ engineering. Terms are organized by domain \u2014 frontend, backend, devops,\
    \ and docs-as-code \u2014 and can be retrieved, filtered, and created via this\
    \ API.\n\n**Base URL:** `https://tech-terms-api-production.up.railway.app`\n\n\
    **Authentication:** None required. All endpoints are publicly accessible."
  contact:
    url: https://github.com/dave-lewis-van/tech-terms/issues
  version: 1.0.0
paths:
  /terms:
    get:
      tags:
      - Terms
      summary: Get Terms
      description: Retrieve a list of technical terms with optional filtering.
      operationId: get_terms_terms_get
      parameters:
      - name: category
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/TermCategory'
        description: Filter terms by technical domain.
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search term names by keyword (case-insensitive).
          title: Search
        description: Search term names by keyword (case-insensitive).
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GlossaryTerm'
                title: Response Get Terms Terms Get
              example:
              - id: 1
                term: SSG
                definition: Static Site Generator.
                category: docs-as-code
                see_also: []
              - id: 101
                term: Hydration
                definition: The process of attaching event listeners to static HTML.
                category: frontend
                see_also:
                - 105
                - 202
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Terms
      summary: Create Term
      description: Register a new technical term in the glossary.
      operationId: create_term_terms_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlossaryTerm'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlossaryTerm'
              example:
                id: 42
                term: Tree Shaking
                definition: The elimination of dead code from a JavaScript bundle
                  at build time.
                category: frontend
                see_also:
                - 101
        '400':
          description: Malformed JSON body
          content:
            application/json:
              example:
                detail: There was an error parsing the body
        '422':
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - category
                  msg: 'value is not a valid enumeration member; permitted: ''frontend'',
                    ''backend'', ''devops'', ''docs-as-code'''
                  type: type_error.enum
                  input: cloud
          description: Unprocessable Entity
  /terms/bulk:
    post:
      tags:
      - Terms
      summary: Bulk Import Terms
      description: Import multiple terms in a single request, returning a per-item
        result for each entry.
      operationId: bulk_import_terms_terms_bulk_post
      requestBody:
        content:
          application/json:
            schema:
              items: {}
              type: array
              title: Items
              minItems: 1
        required: true
      responses:
        '207':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BulkImportResult'
                type: array
                title: Response Bulk Import Terms Terms Bulk Post
              example:
              - index: 0
                status: created
                id: 11
              - index: 1
                status: error
                detail: Term with ID 1 already exists.
        '400':
          description: Empty request body
          content:
            application/json:
              example:
                detail: Request body must contain at least one term.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /terms/{id}:
    get:
      tags:
      - Terms
      summary: Get Term
      description: Fetch a specific term by its ID.
      operationId: get_term_terms__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          description: The unique numeric ID of the glossary term.
          title: Id
        description: The unique numeric ID of the glossary term.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlossaryTerm'
              example:
                id: 101
                term: Hydration
                definition: The process of attaching event listeners to static HTML.
                category: frontend
                see_also:
                - 105
                - 202
        '404':
          description: Term not found
          content:
            application/json:
              example:
                detail: Term not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BulkImportResult:
      properties:
        index:
          type: integer
          title: Index
        status:
          type: string
          title: Status
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
      type: object
      required:
      - index
      - status
      title: BulkImportResult
    GlossaryTerm:
      properties:
        id:
          type: integer
          title: Id
          example: 101
        term:
          type: string
          title: Term
          description: The canonical name of the technical concept.
          example: Hydration
        definition:
          type: string
          title: Definition
          description: A concise, technical explanation of the term.
          example: The process of attaching event listeners to static HTML.
        category:
          $ref: '#/components/schemas/TermCategory'
          example: frontend
        see_also:
          anyOf:
          - items:
              type: integer
            type: array
          - type: 'null'
          title: See Also
          description: IDs of related glossary entries.
          example:
          - 105
          - 202
      type: object
      required:
      - id
      - term
      - definition
      - category
      title: GlossaryTerm
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    TermCategory:
      type: string
      enum:
      - frontend
      - backend
      - devops
      - docs-as-code
      title: TermCategory
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
tags:
- name: Terms
  description: Retrieve and create technical glossary terms.
servers:
- url: https://tech-terms-api-production.up.railway.app
  description: Production
x-readme:
  samples-languages:
  - curl
  - javascript
  - python